requirements.txt
# Core AI/LLM
anthropic>=0.40.0
openai
cohere
google-genai>=1.0.0

# Google Services
google-auth>=2.36.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.160.0
gspread>=6.1.0

# Vector DB & RAG
pinecone-client

# Web Scraping & Automation
apify-client>=1.8.0
beautifulsoup4>=4.12.0
playwright>=1.49.0
httpx>=0.28.0
requests>=2.32.0
html2text>=2024.2.0

# Data Processing
pandas>=2.2.0
Pillow>=11.0.0
numpy>=1.26.0

# Computer Vision (for thumbnail face matching)
mediapipe>=0.10.0
opencv-python>=4.9.0

# Infrastructure
modal>=0.73.0
python-dotenv>=1.0.0

# PDF Processing
pymupdf>=1.24.0
